requirements.txt
pytest==7.4.4
pytest-html==4.1.1
pyyaml==6.0.1
playwright==1.49.0
pytest-playwright==0.4.3
setuptools<70.0.0
